Send empty directories as well
